dividiendo ficheros
